requirements.txt
asgiref==3.3.4
boto3==1.17.57
botocore==1.20.57
dj-database-url==0.5.0
Django==3.2
django-crispy-forms==1.11.2
django-resized==0.3.11
django-storages==1.11.1
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.2.0
psycopg2==2.8.6
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
s3transfer==0.4.2
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
